+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.4.4 (Next)
 
+* [#317](https://github.com/intridea/hashie/pull/317): Ensure `Hashie::Extensions::MethodQuery` methods return boolean values - [@michaelherold](https://github.com/michaelherold).
 * Your contribution here.
 
 ## 3.4.3 (10/25/2015)